+++ resources/views/components/zoomviewer.blade.php
@@ -6,7 +6,7 @@
     <a id="zoom-out" href="#zoom-out" title="zoom out"><i class="fa fa-minus"></i></a>
     <a id="home" href="#home" title="zoom to fit"><i class="fa fa-home"></i></a> 
     <a id="full-page" href="#full-page" title="zobraz fullscreen"><i class="fa fa-expand"></i></a> 
-    @if ($related_items)
+    @if (count($images) > 1)
       <a id="previous" href="#previous" title="predchádzajúce súvisiace dielo"><i class="fa fa-arrow-up"></i></a> 
       <a id="next" href="#next" title="nasledujúce súvisiace dielo"><i class="fa fa-arrow-down"></i></a> 
     @endif
@@ -16,8 +16,8 @@
     <a class="btn btn-default btn-outline return" href="{!! $item->getUrl() !!}" role="button"><i class="fa fa-arrow-left"></i> {{ trans('general.back') }}</a>
   @endif
 
-  @if ($related_items)
-    <div class="autohide"><div class="currentpage"><span id="index">{!! array_search($item->iipimg_url, $related_items ) + 1 !!}</span> / {!! count($related_items) !!}</div></div>
+  @if (count($images) > 1)
+    <div class="autohide"><div class="currentpage"><span id="index">{!! $index + 1 !!}</span> / {!! count($images) !!}</div></div>
   @endif
 
   <div class="credit">