+++ app/Item.php
@@ -160,6 +160,11 @@ class Item extends Model
         return $this->hasOne(\App\SpiceHarvesterRecord::class, 'item_id');
     }
 
+    public function images()
+    {
+        return $this->hasMany(Image::class)->orderBy('order');
+    }
+
     public function getImagePath($full = false)
     {
         return self::getImagePathForId($this->id, $full);
@@ -657,6 +662,14 @@ class Item extends Model
         return $query->where('gallery', '=', 'Slovenská národná galéria, SNG');
     }
 
+
+    public function scopeRelated($query, Item $item)
+    {
+        return $query->where('related_work', '=', $item->related_work)
+            ->where('author', '=', $item->author)
+            ->orderBy('related_work_order');
+    }
+
     public function download()
     {
 
@@ -731,11 +744,23 @@ class Item extends Model
         return implode(', ', $this->authors)  . $dash .  $this->title;
     }
 
+    public function getHasIipAttribute() {
+        return !$this->getZoomableImages()->isEmpty();
+    }
+
+    public function getZoomableImages()
+    {
+        return $this->images->filter(function (Image $image) {
+            return $image->iipimg_url !== null;
+        });
+    }
+
     public function index()
     {
             $client =  $this->getElasticClient();
             $work_types = $this->work_types;
             $main_work_type = reset($work_types);
+
             $data = [
                 'id' => $this->attributes['id'],
                 'identifier' => $this->attributes['identifier'],
@@ -756,7 +781,7 @@ class Item extends Model
                 'updated_at' => $this->attributes['updated_at'],
                 'created_at' => $this->attributes['created_at'],
                 'has_image' => (bool)$this->has_image,
-                'has_iip' => (bool)$this->iipimg_url,
+                'has_iip' => (bool)$this->has_iip,
                 'is_free' => $this->isFree(),
                 // 'free_download' => $this->isFreeDownload(), // staci zapnut is_free + has_iip
                 'related_work' => $this->related_work,
@@ -769,7 +794,7 @@ class Item extends Model
                 'index' => Config::get('bouncy.index'),
                 'type' =>  self::ES_TYPE,
                 'id' => $this->attributes['id'],
-                'body' =>$data,
+                'body' => $data,
             ]);
     }