+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file[^1].
 ### Changed
 - Favicons & Open Graph tags include components
 - positioned image reference strip on right side of zoom-viewer
+- Zoom view showing multiple images
 
 ### Added
 - zoom-viewer component
 - Moravian gallery importer
+- Multiple images per artwork
 
 ## [1.3.1] - 2018-03-09
 ### Fixed